Fixing all the links
Everywhere used absolute path
