+++ branches.php
@@ -1,35 +1,35 @@
 <!DOCTYPE html>
 <html lang="en">
-    <head>
-        <?php 
-            require_once '../config.php'; 
-            require_once './assets/html/head.php';   
-        ?>
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
-        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
-        <title>Branches</title>
-        <link rel="stylesheet" href="./css/style.css">
-        <link rel="stylesheet" href="./css/footer.css">
-        <link rel="stylesheet" href="./css/nav.css">
-    </head>
-    <body>
+
+<head>
     <?php
-        require_once './assets/html/navbar.php';
-        $array = array(
-            // (Button name , Button link , Button type)
-            // class types: "normal", "special", "disabled",
-            'title' => 'van Franken Car Service',
-            'nav' => array(
-                array('Home', './', 'normal'),
-                array('How does it work?', './howdoesitwork.php', 'normal'),
-                array('Prices', './prices.php', 'normal'),
-                array('Contact', './contact.php', 'normal'),
-                array('Sign in', './login.php', 'normal'),
-                array('Schedule your appointment', './order', 'special'),
-            ),
-        );
-        createnavbar($array);
+    require_once './assets/html/head.php';
+    ?>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+    <title>Branches</title>
+    <link rel="stylesheet" href="./css/style.css">
+    <link rel="stylesheet" href="./css/footer.css">
+    <link rel="stylesheet" href="./css/nav.css">
+</head>
+
+<body>
+    <?php
+    $array = array(
+        // (Button name , Button link , Button type)
+        // class types: "normal", "special", "disabled",
+        'title' => 'van Franken Car Service',
+        'nav' => array(
+            array('Home', './', 'normal'),
+            array('How does it work?', './howdoesitwork.php', 'normal'),
+            array('Prices', './prices.php', 'normal'),
+            array('Contact', './contact.php', 'normal'),
+            array('Sign in', './login.php', 'normal'),
+            array('Schedule your appointment', './order', 'special'),
+        ),
+    );
+    createnavbar($array);
     ?>
     <div class="container" style="margin-top: 100px">
         <div class="container">
@@ -45,15 +45,12 @@
         </div>
     </div>
 
-
-
-
     <div class="container">
-        <div class="row gutters-4 gutters-lg-6 align-items-center" >
+        <div class="row gutters-4 gutters-lg-6 align-items-center">
             <div class="col-12 col-lg-7">
                 <h2 class="h3 header">Branches</h2>
                 <p class="container__text">
-                lol why u visiting this page
+                    lol why u visiting this page
                 </p>
                 <a class="btn btn-primary" href="./contact.php" role="button">Just go back...</a>
 
@@ -63,19 +60,15 @@
 
 
     <footer>
-        <?php 
+        <?php
         require_once './assets/html/footer.php';
         ?>
     </footer>
-    </body>
-    <script>
+</body>
+<script>
+    <?php
+    require_once './assets/js/nav.js';
+    ?>
+</script>
 
-        <?php
-        require_once './assets/js/nav.js';
-        ?>
-
-    </script>
 </html>
-
-
-
