+++ customer/dashboard/index.php
@@ -1,0 +1,3 @@
+<!-- Hier komen verschillende dashboards voor alle gebruikers. Dit zijn receptionisten, klanten en alle andere werknemers (ook voor de owner) -->`
+<?php
+echo "Hello World";
