+++ login.php
@@ -1,35 +1,49 @@
+<?php
+
+
+
+?>
+
 <!DOCTYPE html>
 <html>
-    <head>
-        <?php 
-            require_once '../config.php'; 
-            require_once './assets/html/head.php'; 
-        ?>
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
-        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
-        <title>Sign in</title>
-        <link rel="stylesheet" href="./css/style.css">
-        <link rel="stylesheet" href="./css/footer.css">
-        <link rel="stylesheet" href="./css/nav.css">
-    </head>
-    <body>
+
+<head>
     <?php
-        require_once './assets/html/navbar.php';
-        $array = array(
-            // (Button name , Button link , Button type)
-            // class types: "normal", "special", "disabled",
-            'title' => 'van Franken Car Service',
-            'nav' => array(
-                array('Home', './', 'normal'),
-                array('How does it work?', './howdoesitwork.php', 'normal'),
-                array('Prices', './prices.php', 'normal'),
-                array('Contact', './contact.php', 'normal'),
-                array('Sign in', '#', 'normal'),
-                array('Schedule your appointment', './order', 'special'),
-            ),
-        );
-        createnavbar($array);
+    require_once './assets/html/head.php';
+    ?>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+    <title>Sign in</title>
+    <link rel="stylesheet" href="./css/style.css">
+    <link rel="stylesheet" href="./css/footer.css">
+    <link rel="stylesheet" href="./css/nav.css">
+</head>
+
+<body>
+    <?php
+    // Login the pagina
+    if (isset($_GET['action'])) {
+        if ($_GET['action'] == 'login') {
+            try_login();
+        }
+    }
+
+
+    $array = array(
+        // (Button name , Button link , Button type)
+        // class types: "normal", "special", "disabled",
+        'title' => 'van Franken Car Service',
+        'nav' => array(
+            array('Home', './', 'normal'),
+            array('How does it work?', './howdoesitwork.php', 'normal'),
+            array('Prices', './prices.php', 'normal'),
+            array('Contact', './contact.php', 'normal'),
+            array('Sign in', '#', 'normal'),
+            array('Schedule your appointment', './order', 'special'),
+        ),
+    );
+    createnavbar($array);
     ?>
     <div class="container" style="margin-top: 100px">
         <div class="container">
@@ -45,50 +59,48 @@
     </div>
 
 
-        <div class="container">
-            <div class="row gutters-4 gutters-lg-6 align-items-center" >
-                <div class="col-12 col-lg-9">
-                    <?php
-                        require_once './assets/html/alert.php';
-                        alert();
-                    ?>
-                    <h2 class="h3 header">
+    <div class="container">
+        <div class="row gutters-4 gutters-lg-6 align-items-center">
+            <div class="col-12 col-lg-9">
+                <?php
+                alert_session();
+                ?>
+                <h2 class="h3 header">
                     Sign in
-                    </h2>
-                    <p class="container__text">
+                </h2>
+                <p class="container__text">
                     Sign in here with your created account. Don't have an account yet? Then you must make an appointment. <a href="./werkplaatsplanner/1.voertuig/">Click here to make an appointment</a>
-                    </p>
-                    <form class = "login" action="./assets/actions/action.php?action=login" id='login' method="post">
-                        <div class="form-row" style="max-width: 300px; margin-bottom: 10px;" >
-                                <label for="email">Email</label>
-                                <input type="email" class="form-control" id="email" name='email' placeholder="Email">
-                        </div>
+                </p>
+                <form class="login" action="login.php?action=login" id='login' method="post">
+                    <div class="form-row" style="max-width: 300px; margin-bottom: 10px;">
+                        <label for="email">Email</label>
+                        <input type="email" class="form-control" id="email" name='email' placeholder="Email">
+                    </div>
 
-                        <div class="form-row" style="max-width: 300px; margin-bottom: 10px;">
-                                <label for="inputEmail4">Wachtwoord</label>
-                                <input type="password" class="form-control" name='password' id="password" placeholder="Password">
-                        </div>
-                        <div class="form-row">
-                            <button type="submit" class="btn btn-primary">Sign In</button>
-                        </div>
-                    </form>
-                </div>
+                    <div class="form-row" style="max-width: 300px; margin-bottom: 10px;">
+                        <label for="inputEmail4">Wachtwoord</label>
+                        <input type="password" class="form-control" name='password' id="password" placeholder="Password">
+                    </div>
+                    <div class="form-row">
+                        <button type="submit" class="btn btn-primary">Sign In</button>
+                    </div>
+                </form>
             </div>
         </div>
+    </div>
 
 
-        <footer>
-            <?php 
-            require_once './assets/html/footer.php';
-            ?>
-        </footer>
-
-    </body>
-    <script>
-
+    <footer>
         <?php
-        require_once './assets/js/nav.js';
+        require_once './assets/html/footer.php';
         ?>
+    </footer>
 
-    </script>
+</body>
+<script>
+    <?php
+    require_once './assets/js/nav.js';
+    ?>
+</script>
+
 </html>
